requirements.txt
langchain
streamlit
python-dotenv
openai
deeplake
tiktoken
pandasai
tiktoken
scikit-learn
tenacity
rapidfuzz
aiomultiprocess
pydantic